Initial commit
This relates to concourse/concourse#5803.

The README is pretty heinous. We were just at the point of deciding
how best to implement the DB Accountant. The first step seemed to
involve a strategy pattern - one strategy per container owner type
(image pull, resource check, build, etc).

Ideally we would have a "batch" implementation that can do all the
required accounting with a minimum of DB queries, but writing a lot of
raw SQL against a schema we do not control felt risky. We were
considering using the interfaces in
`github.com/concourse/concourse/atc/db` to make a
potentially-inefficient but more type-safe/future-proof implementation
first, and then using that implementation as an oracle for a more
efficient but riskier "batch" version.

Signed-off-by: James Thomson <jthomson@pivotal.io>
Co-authored-by: Jamie Klassen <cklassen@pivotal.io>