create and destroy databases
instead of database server processes. This means that the lifecycle of the
postgres server itself is completely outside the control of the test runner, and
actually means that as long as localhost:5432 is appropriately port-forwarded
without SSL and `postgres` is a superuser you can run your postgres anywhere.

the example we'd like to move toward is following the pattern of
https://github.com/meAmidos/dcind and tracking our dependence on postgres via
an image.

Signed-off-by: Jamie Klassen <jklassen@vmware.com>