WIP: adding k8s worker
struggling with the writing of tests -- initially it had felt realistic to
construct a `K8sWorker` using a `Factory` from `kubectl/pkg/cmd/util` since
`kubectl` already included some helpful tools for creating a test double for
that interface.

However, we are going under the hood when it comes to the port-forwarding
mechanism in such a way that the test clients generated by the test factory
are generating SPDY dialers that attempt to connect to localhost on port 80
rather than returning stubbed responses.

I considered that this might just be a testing problem -- as long as the
implementation was still clean, it might have been worth moving forward with the
same design.

However, it also seems to be non-trivially difficult to reproduce the behaviour
of "hitting the default context specified in your kubeconfig" in the manner that
`kubectl` does -- I have seen some cryptic "no Auth Provider for 'gcp'"-type
message when I try to `go run main.go`.

So I think it's worth revisiting the design.

Our code actually depends more directly on the RESTConfig (to establish
connections) and RESTClient (to determine the structure of the API endpoints)
types -- it may make more sense to inject these types directly. Hopefully that
will allow us to create more useful stubs, and to only extract the
kubeconfig-related data we need without committing to all the boilerplate of
setting up full flag-compatibility with `kubectl`.

concourse/concourse#5954

Signed-off-by: Jamie Klassen <jklassen@vmware.com>